// apps/web/src/lib/api/generated/main/model/create-notification-rule-dto-trigger-type.ts
/**
 * Generated by orval v8.7.0 🍺
 * Do not edit manually.
 * Lootlog API
 * The Lootlog API documentation
 * OpenAPI spec version: 1.0
 */

export type CreateNotificationRuleDtoTriggerType = typeof CreateNotificationRuleDtoTriggerType[keyof typeof CreateNotificationRuleDtoTriggerType];


export const CreateNotificationRuleDtoTriggerType = {
  TIMER_BEFORE_SPAWN: 'TIMER_BEFORE_SPAWN',
  NPC_SPAWNED: 'NPC_SPAWNED',
  WATCHED_ITEM_DROPPED: 'WATCHED_ITEM_DROPPED',
  SCHEDULED_MESSAGE: 'SCHEDULED_MESSAGE',
} as const;